[MOV] core: move RTC verification helpers
