+++ crates/core/src/engine/media_transport/rtc/TESTS/packet_loop_state_tests.rs
@@ -2,7 +2,7 @@ use str0m::media::Rid;
 
 use super::{
     super::{
-        bootstrap::ensure_session_rtc_state,
+        bootstrap::test_support::ensure_session_rtc_state,
         state::{
             PacketLoopState,
             bitrate::{BitrateRegistry, IncomingBitrateObservation},
