+++ crates/core/src/engine/media_transport/rtc/worker/mod.rs
@@ -48,6 +48,10 @@ pub(super) mod session_drain;
 #[expect(non_snake_case, reason = "test modules map to local TESTS directories")]
 mod TESTS;
 
+#[cfg(any(test, feature = "internal-benchmarks"))]
+#[path = "TESTS/commands.rs"]
+mod command_support;
+
 #[cfg(any(test, feature = "testing-transport"))]
 #[path = "TESTS/support.rs"]
 mod test_support;
@@ -58,17 +62,9 @@ use std::{
     thread,
 };
 
-#[cfg(any(test, feature = "internal-benchmarks"))]
-use o_sfu_router::rtp::MediaStream as RouterRtpParameters;
-#[cfg(any(test, feature = "internal-benchmarks"))]
-use str0m::media::MediaKind;
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
-#[cfg(test)]
-use super::commands::ParsedSessionAnswer;
-#[cfg(any(test, feature = "internal-benchmarks"))]
-use super::commands::RtcSessionOffer;
 use super::{
     commands::{RemoteSourceControl, RouteControlRequest, RtcWorkerCommand},
     state::{
@@ -77,12 +73,8 @@ use super::{
         relay_registry::{RelayPacketMailbox, RelayTargetId},
     },
 };
-#[cfg(any(test, feature = "internal-benchmarks"))]
-use crate::engine::media_transport::TransportAdapterError;
 #[cfg(test)]
-use crate::engine::media_transport::{AppliedSessionAnswer, SourcePolicySignal};
-#[cfg(any(test, feature = "internal-benchmarks"))]
-use crate::engine::media_transport::{SessionOffer, TransportMediaId, TransportSessionKey};
+use crate::engine::media_transport::SourcePolicySignal;
 #[cfg(any(test, feature = "testing-transport"))]
 use crate::engine::metrics::RuntimeMetrics;
 use crate::engine::{
@@ -127,131 +119,6 @@ pub struct RtcWorker {
 }
 
 impl RtcWorker {
-    #[cfg(any(test, feature = "internal-benchmarks"))]
-    pub async fn create_initial_session_offer(
-        &self,
-        room_id: &str,
-        session_key: &TransportSessionKey,
-    ) -> Result<SessionOffer, TransportAdapterError> {
-        let room_id: Arc<str> = Arc::from(room_id);
-        self.request_worker(
-            move |response| RtcWorkerCommand::CreateInitialSessionOffer {
-                room_id,
-                session_key: session_key.clone(),
-                response,
-            },
-        )
-        .await
-        .map(RtcSessionOffer::into_session_offer)
-    }
-
-    #[cfg(test)]
-    pub async fn create_session_renegotiation_offer(
-        &self,
-        session_key: &TransportSessionKey,
-    ) -> Result<SessionOffer, TransportAdapterError> {
-        self.request_worker(
-            |response| RtcWorkerCommand::CreateSessionRenegotiationOffer {
-                session_key: session_key.clone(),
-                response,
-            },
-        )
-        .await
-        .map(RtcSessionOffer::into_session_offer)
-    }
-
-    #[cfg(test)]
-    pub async fn apply_session_answer(
-        &self,
-        session_key: &TransportSessionKey,
-        answer_sdp: &str,
-    ) -> Result<AppliedSessionAnswer, TransportAdapterError> {
-        let answer = ParsedSessionAnswer::parse(answer_sdp)?;
-        self.request_worker(|response| RtcWorkerCommand::ApplySessionAnswer {
-            session_key: session_key.clone(),
-            answer,
-            response,
-        })
-        .await
-    }
-    #[cfg(any(test, feature = "internal-benchmarks"))]
-    pub async fn close_session(
-        &self,
-        session_key: &TransportSessionKey,
-    ) -> Result<(), TransportAdapterError> {
-        self.request_worker(|response| RtcWorkerCommand::CloseSession {
-            session_key: session_key.clone(),
-            response,
-        })
-        .await
-    }
-    #[cfg(any(test, feature = "internal-benchmarks"))]
-    pub async fn remove_media(
-        &self,
-        session_key: &TransportSessionKey,
-        transport_media_id: TransportMediaId,
-    ) -> Result<(), TransportAdapterError> {
-        self.request_worker(|response| RtcWorkerCommand::RemoveMedia {
-            session_key: session_key.clone(),
-            transport_media_id,
-            response,
-        })
-        .await
-    }
-
-    #[cfg(test)]
-    pub async fn negotiated_producer_parameters(
-        &self,
-        session_key: &TransportSessionKey,
-        transport_media_id: TransportMediaId,
-    ) -> Result<RouterRtpParameters, TransportAdapterError> {
-        self.request_worker(
-            |response| RtcWorkerCommand::ResolveNegotiatedProducerParameters {
-                session_key: session_key.clone(),
-                transport_media_id,
-                response,
-            },
-        )
-        .await
-    }
-
-    #[cfg(any(test, feature = "internal-benchmarks"))]
-    pub async fn add_recv_media(
-        &self,
-        session_key: &TransportSessionKey,
-        media_kind: MediaKind,
-        rtp_parameters: &RouterRtpParameters,
-    ) -> Result<TransportMediaId, TransportAdapterError> {
-        self.request_worker(|response| RtcWorkerCommand::AddRecvMedia {
-            session_key: session_key.clone(),
-            media_kind,
-            rtp_parameters: rtp_parameters.clone(),
-            response,
-        })
-        .await
-    }
-
-    #[cfg(test)]
-    pub async fn add_send_media(
-        &self,
-        consumer_key: &TransportSessionKey,
-        media_kind: MediaKind,
-        source: TransportSourceKey,
-        consumer_rtp_parameters: &RouterRtpParameters,
-        active: bool,
-    ) -> Result<TransportMediaId, TransportAdapterError> {
-        self.request_worker(|response| RtcWorkerCommand::AddSendMedia {
-            consumer_key: consumer_key.clone(),
-            media_kind,
-            source,
-            remote_source_control: None,
-            consumer_rtp_parameters: consumer_rtp_parameters.clone(),
-            active,
-            response,
-        })
-        .await
-    }
-
     /// Builds the handle `consumer` stores for a producer owned by `self`.
     ///
     /// The returned handle lets the consumer worker send best-effort keyframe
